backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
pypdf
python-docx
python-multipart
supabase
python-jose
